Extract packages for handling external tools
